Fixed bug with pointers where a pointer type would be converted to what it points to. (Sprite would become SpriteData). HowTo packages are generated properly and archived. Fixed an issue with the C include statements not having .h at the end of them. Fixed some spacing issues in the templates.

+++ Tools/HowToPackager/package_how_tos.py
@@ -14,9 +14,10 @@ import os
 import shutil
 from shutil import make_archive
 
-_lang = [   {"lang": "Pascal", "template": "../../Dist/Pascal/FPC", "main file": "GameMain.pas", "extension": ".pas"}, 
+_lang = [   
+            {"lang": "Pascal", "template": "../../Dist/Pascal/FPC", "main file": "GameMain.pas", "extension": ".pas"}, 
             {"lang": "C", "template": "../../Dist/C/gcc", "main file": "main.c", "extension": ".c"} 
-            ]
+        ]
 _base_path = "../../Dist/HowTo"
 
 def build_dir_structure():
@@ -47,23 +48,33 @@ def copy_how_to_template(how_to_name):
 def copy_how_to_resources(current_how_to, how_to_resource, lang):
     path = "%s/%s" % ("../../CoreSDK/test/Resources",how_to_resource)
     dest = "%s/%s/%s/%s/%s" % (_base_path, lang["lang"],current_how_to,"Resources",how_to_resource)
-    print "   -- Copying ", path
-    print "   -- To ",dest        
-    shutil.copy2(path,dest)
+    if os.path.exists(path):
+        print "   -- Copying ", path
+        print "   -- To ",dest        
+        shutil.copy2(path,dest)
+    else:
+        print "   WARNING -- "+ path +" file does not exist"
 
 def zip_how_to():
     f = open("../../CoreSDK/test/HowToResources.txt")
     how_to_file_lines = f.readlines()
     f.close()
     for line in how_to_file_lines:
-           # if line starts with *... its a new how to...
-           if line[0] == "*":
-               current_how_to = line[1:].strip()
-               for lang in _lang:
-                    root_dir = os.path.expanduser(os.path.join(_base_path, lang["lang"], current_how_to))
+       # if line starts with *... its a new how to...
+       if line[0] == "*":
+           current_how_to = line[1:].strip()
+           for lang in _lang:
+                root_dir = os.path.expanduser(os.path.join(_base_path, lang["lang"], current_how_to))
+                how_to_src = root_dir + "/src/" + lang['main file']
+                print " - Checking How To src file: %s" % how_to_src
+                if os.path.exists(how_to_src):
                     path = "%s/%s/%s/%s" % (_base_path, lang["lang"], "Archive", current_how_to)
                     print "   -- Creating Archive "
+                    print "     - ", current_how_to
                     make_archive(path, 'zip',root_dir)
+                else:
+                    print " - Skipping: %s" % how_to_src
+                    print "     - Unable to find source file"
 
 def copy_current_how_to_pas_file():
     f = open("../../CoreSDK/test/HowToResources.txt")
@@ -74,17 +85,14 @@ def copy_current_how_to_pas_file():
         if line[0] == "*":
             current_how_to = line[1:].strip()
             for lang in _lang:
-                path = "%s/%s%s" % (_base_path + "Source_Code/" + lang['lang'] + '/',current_how_to,lang['extension'])
-                if os.path.exists(path):                    
-                    dest = "%s/%s/%s/%s/%s%s" % (_base_path, lang["lang"],current_how_to,"src",current_how_to, lang['extension'])
+                path = "%s/%s%s" % (_base_path + "/Source_Code/" + lang['lang'] + '/',current_how_to,lang['extension'])
+                if os.path.exists(path):            
+                    # copy the current how to file into the src directory        
+                    dest = "%s/%s/%s/%s/%s" % (_base_path, lang["lang"],current_how_to,"src",lang['main file'])
+                    print " - Copying GameMain"
                     print "   -- Copying ", path
                     print "   -- To ",dest        
                     shutil.copy2(path,dest)
-                    sourcePath = "%s/%s%s" % ("../../CoreSDK/test",current_how_to,".pas")
-                    sourceDest = "%s/%s/%s%s" % (_base_path,"Source_Code", current_how_to,".pas")
-                    print "   -- Copying ", sourcePath
-                    print "   -- To ",sourceDest        
-                    shutil.copy2(sourcePath,sourceDest)
                 else:
                     print "   WARNING -- "+ current_how_to +" file does not exist"
